+++ .gitattributes
@@ -1,1 +1,2 @@
 **/testData/** -linguist-detectable
+howdju-text-fragment-generation/dist/** -linguist-detectable
