+++ howdju-common/lib/urlTextFragments.ts
@@ -1,8 +1,12 @@
 import * as textQuote from "dom-anchor-text-quote";
-import * as textPosition from "dom-anchor-text-position";
 
-import { isDefined, logger, nodeIsAfter, UrlLocator } from "howdju-common";
-import { indexOf } from "lodash";
+import {
+  getTextWithin,
+  isDefined,
+  logger,
+  toPlainTextContent,
+  UrlLocator,
+} from "howdju-common";
 
 const FRAGMENT_DIRECTIVE = ":~:";
 
@@ -166,204 +170,6 @@ export function extractQuotationFromTextFragment(
     return textParameters.slice(start, end + 1).join(textDirectiveDelimiter);
   });
   return quoteParts.filter(isDefined).join(options.textDirectiveDelimiter);
-}
-
-function getTextWithin(doc: Document, startText: string, endText: string) {
-  // Some sites includes the content of the page in a script tag. E.g. substack's `body_html`. So
-  // use a hint at the beginning to try and find content in the body. (If we find this doens't work,
-  // we might need to use a binary search style approach until we either have exhausted ranges in
-  // the document or have found a range that isn't in a script tag.)
-  const { range } = getRangeOfText(doc, startText, endText, 0);
-  if (!range) {
-    return undefined;
-  }
-  if (range.collapsed) {
-    return undefined;
-  }
-  return toPlainTextContent(range);
-}
-
-function getRangeOfText(
-  doc: Document,
-  startText: string,
-  endText: string,
-  hint?: number
-) {
-  let startPosition = textQuote.toTextPosition(
-    doc.body,
-    { exact: startText },
-    hint !== undefined ? { hint } : undefined
-  );
-  if (!startPosition) {
-    return { range: undefined, end: undefined };
-  }
-  let endPosition = textQuote.toTextPosition(
-    doc.body,
-    { exact: endText },
-    { hint: startPosition.end }
-  );
-  if (!endPosition) {
-    return { range: undefined, end: undefined };
-  }
-  // If the positions are invalid, try to find better positions.
-  if (startPosition.start >= endPosition.end) {
-    const betterStartPosition = textQuote.toTextPosition(
-      doc.body,
-      { exact: startText },
-      { hint: endPosition.start }
-    );
-    const betterEndPosition = textQuote.toTextPosition(
-      doc.body,
-      { exact: endText },
-      { hint: startPosition.end }
-    );
-    const betterStartLength = betterStartPosition
-      ? endPosition.start - betterStartPosition.end
-      : Number.NEGATIVE_INFINITY;
-    const betterEndLength = betterEndPosition
-      ? betterEndPosition.start - startPosition.end
-      : Number.NEGATIVE_INFINITY;
-    const isValidBetterStart = betterStartPosition && betterStartLength > 0;
-    const isValidBetterEnd = betterEndPosition && betterEndLength > 0;
-    if (isValidBetterStart) {
-      if (isValidBetterEnd) {
-        // If both better positions were found, return the one that yields a smaller range.
-        if (betterStartLength < betterEndLength) {
-          startPosition = betterStartPosition;
-        } else {
-          endPosition = betterEndPosition;
-        }
-      } else {
-        startPosition = betterStartPosition;
-      }
-    } else if (isValidBetterEnd) {
-      endPosition = betterEndPosition;
-    }
-    // If the positions are still invalid, give up.
-    if (startPosition.start >= endPosition.end) {
-      return { range: undefined, end: undefined };
-    }
-  }
-
-  const range = textPosition.toRange(doc.body, {
-    start: startPosition.start,
-    end: endPosition.end,
-  });
-  return { range, end: endPosition.end };
-}
-
-function getNodeConstructor(node: Node) {
-  const Node = node.ownerDocument?.defaultView?.Node;
-  if (!Node) {
-    throw new Error(`Unable to obtain Node constructor from range.`);
-  }
-  return Node;
-}
-
-/**
- * Try to return the contents of a range formatted as plain text.
- *
- * - Every time we encounter a text node, we add its textContent to the result.
- * - Every time we encounter a paragraph, we add two newlines to the result.
- *
- * JSDOM doesn't implement innerText, so we must do this ourselves
- * (https://github.com/jsdom/jsdom/issues/1245). Another option might be to run headless Chrome.
- *
- * (On the differences between textContent and innerText:
- * https://developer.mozilla.org/en-US/docs/Web/API/Node/textContent#differences_from_innertext)
- */
-function toPlainTextContent(range: Range) {
-  const textParts = [] as string[];
-  const Node = getNodeConstructor(range.startContainer);
-  walkRangeNodes(range, {
-    enter: (node) => {
-      if (node.nodeType === Node.TEXT_NODE) {
-        let text;
-        if (node.isSameNode(range.startContainer)) {
-          if (node.isSameNode(range.endContainer)) {
-            text = node.textContent?.substring(
-              range.startOffset,
-              range.endOffset
-            );
-          } else {
-            text = node.textContent?.substring(range.startOffset);
-          }
-        } else if (node.isSameNode(range.endContainer)) {
-          text = node.textContent?.substring(0, range.endOffset);
-        } else {
-          text = node.textContent;
-        }
-        if (text) {
-          textParts.push(text);
-        }
-      }
-    },
-    leave: (node) => {
-      if (
-        node.nodeType === Node.ELEMENT_NODE &&
-        node.nodeName.toLowerCase() === "p"
-      ) {
-        textParts.push("\n\n");
-      }
-    },
-  });
-  return textParts.join("").replace(/\s+$/gm, "\n").trim();
-}
-
-function isTextNode(node: Node): node is Text {
-  return node.nodeType === getNodeConstructor(node).TEXT_NODE;
-}
-
-export function walkRangeNodes(
-  range: Range,
-  { enter, leave }: { enter: (node: Node) => void; leave: (node: Node) => void }
-) {
-  // If the startContainer is not a text node, then startOffset points to the first node of the
-  // range. If the startOffset is 0, it is ambiguous whether the range starts at the startContainer
-  // or at its first child. We assume that a startOffset of 0 includes the startContainer.
-
-  let node: Node | null =
-    isTextNode(range.startContainer) || range.startOffset == 0
-      ? range.startContainer
-      : range.startContainer.childNodes[range.startOffset];
-  while (node) {
-    enter(node);
-    if (node.firstChild) {
-      node = node.firstChild;
-      continue;
-    }
-    while (node && !node.nextSibling) {
-      leave(node);
-      if (node.isSameNode(range.endContainer)) {
-        return;
-      }
-      node = node.parentNode;
-    }
-    if (!node) {
-      logger.error(
-        `Unexpectedly reached the root node without encountering the range's endContainer.`
-      );
-      return;
-    }
-    leave(node);
-    node = node.nextSibling;
-    // If the endContainer is not a text node, then endOffset points to the last node of the range
-    // and we should skip any nodes after it.
-    if (
-      !isTextNode(range.endContainer) &&
-      node?.parentNode?.isSameNode(range.endContainer) &&
-      indexOf(node.parentNode.childNodes, node) >= range.endOffset
-    ) {
-      return;
-    }
-    if (
-      node &&
-      !range.endContainer.contains(node) &&
-      nodeIsAfter(node, range.endContainer)
-    ) {
-      return;
-    }
-  }
 }
 
 export type QuotationConfirmationResult =
